==> Markdowns/07_DataIntegrationChallengeSolution.Rmd <==
---
title: "DataIntergrationChallenge"
author: "Abbi Edwards"
date: "07/06/2021"
output: html_document
---

```{r setup, include=FALSE}
library(ggplot2)
library(scater)
library(scran)
library(batchelor)
library(bluster)
library(tidyverse)
library(pheatmap)
library(clustree)
library(Cairo)
library(BiocSingular)
library(cowplot)
```

## Loading the data you need


```{r Load}

p.tsne <- readRDS("./Robjects/DataIntegration_p.tsne.Rds")
mnn.out <- readRDS("./Robjects/DataIntegration_mnn.out.Rds")

```

## Challenge 1

Draw the TSNE for the fastmnn corrected data. Compare it with the uncorrected TSNE, what do you think?

```{r yourCode1}

mp.tsne <- plotTSNE(mnn.out, colour_by="SampleName", shape_by = "SampleGroup")
mp.tsne

```

```{r}
mp.tsne + facet_wrap(. ~ mnn.out$SampleGroup)
```

```{r tsne compare}
plot_grid(p.tsne,mp.tsne)
```

```
