==> Markdowns/07_DataSetIntegration_PBMMC_ETV6-RUNX1.Rmd <==
---
title: "CRUK CI Summer School 2020 - introduction to single-cell RNA-seq analysis"
subtitle: 'Data integration'
author: "Stephane Ballereau, Abbi Edwards"
output:
  html_document:
    df_print: paged
    toc: yes
    number_sections: true
    code_folding: hide
---

# Data integration - PBMMC and ETV6-RUNX samples


## Learning Objectives

* Understand why/when batch correction is required
* Understand where data set integration fits in the workflow
* Understand one main method for batch correction
* Understand how to check your batch correction has worked

```{r library_PBMMC_ETV6-RUNX1, warning=FALSE, message=FALSE}
library(scater)
library(scran)
library(batchelor)
library(bluster)
library(tidyverse)
library(pheatmap)
library(clustree)
library(Cairo)
library(BiocSingular)
library(cowplot)
```

Source: ['Integrating Datasets'](https://osca.bioconductor.org/integrating-datasets.html) chapter in the OSCA book. 

## Abbreviations
 
* HVG: highly variable genes
* MNN: mutual nearest neighbors
* PBMMC: peripheral blood mononuclear cell
* SCE: SingleCellExperiment

## Motivation

Large single-cell RNA sequencing (scRNA-seq) projects usually need to generate data across multiple batches due to logistical constraints. However, the processing of different batches is often subject to uncontrollable differences, e.g., changes in operator, differences in reagent quality. This results in systematic differences in the observed expression in cells from different batches, which we refer to as “batch effects”. Batch effects are problematic as they can be major drivers of heterogeneity in the data, masking the relevant biological differences and complicating interpretation of the results.

Computational correction of these effects is critical for eliminating batch-to-batch variation, allowing data across multiple batches to be combined for common downstream analysis. However, existing methods based on linear models (Ritchie et al. 2015; Leek et al. 2012) assume that the composition of cell populations are either known or the same across batches. To overcome these limitations, bespoke methods have been developed for batch correction of single-cell data (Haghverdi et al. 2018; Butler et al. 2018; Lin et al. 2019) that do not require a priori knowledge about the composition of the population. This allows them to be used in workflows for exploratory analyses of scRNA-seq data where such knowledge is usually unavailable.

## Load the data

We will load the SCE R object with the normalised counts from after the dimmensionality reduction. This object has each of the 7 individual samples SCE 
objects in a `large list` for ease of running the following steps but it is no different to running each sample seperately at this point. Each sample has been subset to 500 cells per sample for demonstration purposes.

```{r loadSplit_PBMMC_ETV6-RUNX1}
sce <- readRDS("./Robjects/caron_postDeconv_5hCellPerSpl_dimRed.Rds")

#split.sce <- split(sce, sample(LETTERS, nrow(sce), replace=TRUE))
x <- subset(sce, , SampleName=="ETV6-RUNX1_1")
sampleNameLevels <- levels(factor(colData(sce)$SampleName))
sampleNameToGet <- grep("PBMMC|ETV6-RUNX1", sampleNameLevels, value = TRUE)
all.sce <- lapply(sampleNameToGet, function(x){ subset(sce, , SampleName==x) })
names(all.sce) <- sampleNameToGet

saveRDS(all.sce, "./Robjects/DataIntegration_all_sce.Rds")
```

```{r load_PBMMC_ETV6-RUNX1, eval=FALSE}
all.sce <- readRDS("./Robjects/DataIntegration_all_sce.Rds")
all.sce
```

We then apply the standard workflow to each sample separately:

* [normalisation](./normalisation_5hCellPerSpl_caron.Rmd)
* variance modelling
* [dimensionality reduction](./dimRedFeatSel_2021.Rmd)
* [clustering](./clusteringPostDsi.Rmd) 

```{r standard_workflow_PBMMC_ETV6-RUNX1}
#--- normalization ---#
# use logNormCounts()
all.sce <- lapply(all.sce, logNormCounts)

#--- variance-modelling ---#
# model variance with modelGeneVar()
# find highly variable genes (HVGs) with getTopHVGs()
all.dec <- lapply(all.sce, modelGeneVar)
all.hvgs <- lapply(all.dec, getTopHVGs, prop=0.1)

#--- dimensionality-reduction ---#
# use runPCA()
# then compute embeddings with runTSNE() and runUMAP()
set.seed(10000)
all.sce <- mapply(FUN=runPCA,
		  x=all.sce,
		  subset_row=all.hvgs,
		  MoreArgs=list(ncomponents=25,
				BSPARAM=RandomParam()),
		  SIMPLIFY=FALSE)

set.seed(100000)
all.sce <- lapply(all.sce, runTSNE, dimred="PCA")

set.seed(1000000)
all.sce <- lapply(all.sce, runUMAP, dimred="PCA")

#--- clustering ---#
# cluster each sample separately
for (n in names(all.sce)) {
    g <- buildSNNGraph(all.sce[[n]], k=10, use.dimred='PCA')
    clust <- igraph::cluster_walktrap(g)$membership
    all.sce[[n]]$label  <- factor(clust)
}
```

To prepare for the batch correction:

* We subset all batches to the common “universe” of features. In this case, it is straightforward as all the samples in these experiments use the same Ensembl gene annotation.

```{r prep_universe_PBMMC_ETV6-RUNX1}
allNames <- unlist(lapply(all.sce, function(x){rownames(x)}))
allNamesNb <- table(allNames)
universe <- names(allNamesNb)[allNamesNb==7] # where 7 is number of samples
```

* The size of this common “universe” of features here is the number of features shared by all 7 samples is: `r length(universe)`.

```{r universe_subset_PBMMC_ETV6-RUNX1}
# Subsetting the SingleCellExperiment object.
uni.sce <- lapply(all.sce, function(x){x[universe,]})
# Also subsetting the variance modelling results, for convenience.
uni.dec <- lapply(all.dec, function(x){x[universe,]})
```

* We rescale each batch to adjust for differences in sequencing depth between batches. The multiBatchNorm() function recomputes log-normalized expression values after adjusting the size factors for systematic differences in coverage between SingleCellExperiment (SCE) objects. (Size factors only remove biases between cells within a single batch.) This improves the quality of the correction by removing one aspect of the technical differences between batches.

```{r multiBatchNorm_PBMMC_ETV6-RUNX1}
# rescale each batch to adjust for differences in sequencing depth between batches
rescaled <- multiBatchNorm(uni.sce, batch = "SampleName")
```

* We perform feature selection by averaging the variance components across all batches with the combineVar() function. We compute the average as it is responsive to batch-specific HVGs while still preserving the within-batch ranking of genes.


```{r combineVar_PBMMC_ETV6-RUNX1}
# compute average variance components across samples
combined.dec <- combineVar(uni.dec)

# identify highly variables genes
# here as those with a positive biological component
chosen.hvgs <- combined.dec$bio > 0
combined.dec$chosenHvg <- chosen.hvgs
```

Number of HVGs: `r sum(chosen.hvgs)`.

When integrating datasets of variable composition, it is generally safer to err on the side of including more genes than are used in a single dataset analysis, to ensure that markers are retained for any dataset-specific subpopulations that might be present. For a top X selection, this means using a larger X (say, ~5000), or in this case, we simply take all genes above the trend.

Alternatively, a more forceful approach to feature selection can be used based on marker genes from within-batch comparisons.

## Diagnosing batch effects

Before we actually perform any correction, it is worth examining whether there is any batch effect in this dataset. We combine the SCE objects and perform a PCA on the log-expression values for all genes with positive (average) biological components.

```{r diagPrep_PBMMC_ETV6-RUNX1}
# Reminder: the metadata must be the same for each sample for cbind()ing.
# concat matrices:
uncorrected <- do.call(cbind, rescaled)

# copy choice to uncorrected SCE:
all(rownames(combined.dec) == rownames(uncorrected))
rowData(uncorrected) <- cbind(rowData(uncorrected), combined.dec)

saveRDS(uncorrected, "./Robjects/DataIntegration_uncorrected.Rds")
```


```{r}
# Perform PCA
# Using RandomParam() as it is more efficient for file-backed matrices.
set.seed(0010101010)
uncorrected <- runPCA(uncorrected,
                      subset_row=chosen.hvgs,
                      BSPARAM=BiocSingular::RandomParam())
```

We use graph-based clustering on the components to obtain a summary of the population structure. Different clustering methods are discussed in a later [chapter](./clusteringPostDsi.Rmd).

As the samples should be replicates, each cluster should ideally consist of cells from each batch. However, we instead see clusters that are comprised of cells from a single batch. This indicates that cells of the same type are artificially separated due to technical differences between batches.

```{r diagClu_PBMMC_ETV6-RUNX1}
# build shared nearest-neighbour graph
snn.gr <- buildSNNGraph(uncorrected, use.dimred="PCA")
# identify cluster with the walk trap method
clusters <- igraph::cluster_walktrap(snn.gr)$membership
# get number of cells for each {cluster, batch} pair
clusterTab <- data.frame("clusters" = clusters, "batch" = uncorrected$SampleName)
```

Cluster size and cell contribution by sample:

```{r clusPlotdata, warning=FALSE, message=FALSE}
ClusterInfo <- clusterTab %>% 
  group_by(clusters, batch) %>%
  summarise(cells = n()) 
```

```{r diagPlotDraw_PBMMC_ETV6-RUNX1, fig.height=6, fig.width=7,warning=F}
p1 <- ggplot(data=ClusterInfo, aes(x=clusters,y=cells, fill=batch)) +
	geom_col() +
	theme(axis.text.x=element_text(angle = 90, hjust = 0)) +
  ggtitle("uncorrected, cell numbers") +
  theme(legend.text = element_text(size = 7))
p2 <- ggplot(data=clusterTab, aes(x=clusters, fill = batch)) +
  geom_bar(position = "fill") +
  scale_y_continuous(labels = scales::percent) +
	theme(axis.text.x=element_text(angle = 90, hjust = 0)) +
  ggtitle("uncorrected, proportions") +
  theme(legend.text = element_text(size = 7))

plot_grid(p1, p2, ncol=1)
```

We can also visualize the uncorrected coordinates using a t-SNE plot. The strong separation between cells from different batches is consistent with the clustering results.

```{r diagTsneDraw_PBMMC_ETV6-RUNX1}
set.seed(1111001)
uncorrected <- runTSNE(uncorrected, dimred = "PCA")
# draw:
p.tsne <- plotTSNE(uncorrected,
	      colour_by = "SampleName",
	      shape_by = "SampleGroup") +
theme(legend.text = element_text(size = 7))

p.tsne
```

```{r diagTsnePlotSplit_PBMMC_ETV6-RUNX1, fig.height=4, fig.width=7}
p.tsne + facet_wrap(~uncorrected$SampleGroup)
```

Of course, the other explanation for batch-specific clusters is that there are cell types that are unique to each batch. The degree of intermingling of cells from different batches is not an effective diagnostic when the batches involved might actually contain unique cell subpopulations. If a cluster only contains cells from a single batch, one can always debate whether that is caused by a failure of the correction method or if there is truly a batch-specific subpopulation. For example, do batch-specific metabolic or differentiation states represent distinct subpopulations? Or should they be merged together? We will not attempt to answer this here, only noting that each batch correction algorithm will make different (and possibly inappropriate) decisions on what constitutes “shared” and “unique” populations.

Let us write out the corresponding SCE object.

```{r diagWriteRds_PBMMC_ETV6-RUNX1}
# update DataIntegration_uncorrected.Rds file:
saveRDS(uncorrected, "./Robjects/DataIntegration_uncorrected.Rds")
```

## Linear regression

Batch effects in bulk RNA sequencing studies are commonly removed with linear regression. This involves fitting a linear model to each gene’s expression profile, setting the undesirable batch term to zero and recomputing the observations sans the batch effect, yielding a set of corrected expression values for downstream analyses. Linear modelling is the basis of the `removeBatchEffect()` function from the limma package (Ritchie et al. 2015) as well the `comBat()` function from the sva package (Leek et al. 2012).

To use this approach in a scRNA-seq context, we assume that the composition of cell subpopulations is the same across batches. We also assume that the batch effect is additive, i.e., any batch-induced fold-change in expression is the same across different cell subpopulations for any given gene. These are strong assumptions as batches derived from different individuals will naturally exhibit variation in cell type abundances and expression. Nonetheless, they may be acceptable when dealing with batches that are technical replicates generated from the same population of cells. (In fact, when its assumptions hold, linear regression is the most statistically efficient as it uses information from all cells to compute the common batch vector.) Linear modelling can also accommodate situations where the composition is known a priori by including the cell type as a factor in the linear model, but this situation is even less common.

We use the `rescaleBatches()` function from the `batchelor` package to remove the batch effect. This is roughly equivalent to applying a linear regression to the log-expression values per gene, with some adjustments to improve performance and efficiency. For each gene, the mean expression in each batch is scaled down until it is equal to the lowest mean across all batches. We deliberately choose to scale all expression values down as this mitigates differences in variance when batches lie at different positions on the mean-variance trend. (Specifically, the shrinkage effect of the pseudo-count is greater for smaller counts, suppressing any differences in variance across batches.) An additional feature of `rescaleBatches()` is that it will preserve sparsity in the input matrix for greater efficiency, whereas other methods like `removeBatchEffect()` will always return a dense matrix.

```{r linReg_rescaleBatches_PBMMC_ETV6-RUNX1, warning=FALSE}
linear_rescaled <- rescaleBatches(rescaled)
linear_rescaled
```

After clustering, we observe that most clusters consist of mixtures of cells from the two replicate batches, consistent with the removal of the batch effect. This conclusion is supported by the apparent mixing of cells from different batches in Figure 13.2. However, at least one batch-specific cluster is still present, indicating that the correction is not entirely complete. This is attributable to violation of one of the aforementioned assumptions, even in this simple case involving replicated batches.

```{r linReg_diagClu_PBMMC_ETV6-RUNX1}
set.seed(1010101010) # To ensure reproducibility of IRLBA.
linear_rescaled <- runPCA(linear_rescaled, subset_row = chosen.hvgs, exprs_values = "corrected")

snn.gr.linear <- buildSNNGraph(linear_rescaled, use.dimred = "PCA")
clusters.linear <- igraph::cluster_walktrap(snn.gr.linear)$membership

clusterTab.linear <- data.frame("clusters" = clusters.linear, "batch" = linear_rescaled$batch)
```

Cluster size and cell contribution by sample:

```{r linReg_diagPlotDraw_PBMMC_ETV6-RUNX1, message=FALSE}
ClusterInfo.linear <- clusterTab.linear %>% 
  group_by(clusters,batch) %>%
  summarise(cells = n())
```

```{r, fig.height=6, fig.width=7}
lp1 <- ggplot(data=ClusterInfo.linear, aes(x=clusters,y=cells, fill=batch)) +
	theme(axis.text.x=element_text(angle = 90, hjust = 0)) +
	geom_col() +
	theme(legend.text = element_text(size = 7))
lp2 <- ggplot(data=clusterTab.linear, aes(x=clusters, fill=batch)) +
  geom_bar(position = "fill") +
	theme(axis.text.x=element_text(angle = 90, hjust = 0)) +
  scale_y_continuous(labels = scales::percent) +
  theme(legend.text = element_text(size = 7))
plot_grid(lp1, lp2, ncol=1)
```

Compute and plot t-SNE:

```{r linReg_diagTsneDraw_PBMMC_ETV6-RUNX1, fig.height=7, fig.width=7}
linear_rescaled <- runTSNE(linear_rescaled, dimred = "PCA")
linear_rescaled$batch <- factor(linear_rescaled$batch)
lp.tsne <- plotTSNE(linear_rescaled, colour_by = "batch")
lp.tsne
```

```{r linReg_diagTsnePlotSplit_dsi_PBMMC_ETV6-RUNX1, fig.height=4, fig.width=7}
lp.tsne + facet_wrap(~uncorrected$SampleGroup)
```

## Mutual Nearest Neighbour correction

### Algorithm overview

Consider a cell a in batch A, and identify the cells in batch B that are nearest neighbors to a in the expression space defined by the selected features. Repeat this for a cell b in batch B, identifying its nearest neighbors in A. Mutual nearest neighbors are pairs of cells from different batches that belong in each other’s set of nearest neighbors. The reasoning is that MNN pairs represent cells from the same biological state prior to the application of a batch effect - see Haghverdi et al. (2018) for full theoretical details. Thus, the difference between cells in MNN pairs can be used as an estimate of the batch effect, the subtraction of which yields batch-corrected values.

Compared to linear regression, MNN correction does not assume that the population composition is the same or known beforehand. This is because it learns the shared population structure via identification of MNN pairs and uses this information to obtain an appropriate estimate of the batch effect. Instead, the key assumption of MNN-based approaches is that the batch effect is orthogonal to the biology in high-dimensional expression space. Violations reduce the effectiveness and accuracy of the correction, with the most common case arising from variations in the direction of the batch effect between clusters. Nonetheless, the assumption is usually reasonable as a random vector is very likely to be orthogonal in high-dimensional space.


### Application to the data

The `batchelor` package provides an implementation of the MNN approach via the `fastMNN()` function. (Unlike the MNN method originally described by Haghverdi et al. (2018), the `fastMNN()` function performs PCA to reduce the dimensions beforehand and speed up the downstream neighbor detection steps.) We apply it to our two PBMC batches to remove the batch effect across the highly variable genes in `chosen.hvgs`. To reduce computational work and technical noise, all cells in all batches are projected into the low-dimensional space defined by the top d principal components. Identification of MNNs and calculation of correction vectors are then performed in this low-dimensional space.

```{r fastmnnRun_dsi_PBMMC_ETV6-RUNX1}
# Using randomized SVD here, as this is faster than 
# irlba for file-backed matrices.
set.seed(1000101001)
mnn.out <- fastMNN(rescaled,
                   auto.merge = TRUE,
                   d = 50,
                   k = 20,
                   subset.row = chosen.hvgs,
                   BSPARAM=BiocSingular::RandomParam(deferred=TRUE))
mnn.out
mnn.out.corre.dim <- dim(reducedDim(mnn.out, "corrected"))
mnn.out.recon.dim <- dim(assay(mnn.out, "reconstructed"))

# add feature selection outcome to mmn.out
# used in other analyses later.
columnsToGet <- setdiff(colnames(combined.dec), "per.block")
combined.dec.df <- combined.dec[,columnsToGet] %>%
  data.frame() %>%
  tibble::rownames_to_column("ID") %>%
  filter(ID %in% rownames(rowData(mnn.out)))
rotationMat <- rowData(mnn.out)$rotation
rowData(mnn.out)$ID <- rownames(rowData(mnn.out))
rowData(mnn.out)$rotation <- NULL

rowData(mnn.out) <- rowData(mnn.out) %>%
  data.frame() %>%
  left_join(combined.dec.df, by="ID") %>%
  DataFrame()

# add rotation back
rowData(mnn.out)$rotation <- rotationMat

# also have gene symbol:
# copied from 'uncorrected'
rowData(mnn.out)$Symbol <- rowData(uncorrected)[rowData(mnn.out)$ID,]$Symbol

# tidy
rm(columnsToGet, rotationMat)
```

The function returns a SCE object containing corrected values for downstream analyses like clustering or visualization. Each column of `mnn.out` corresponds to a cell in one of the batches, while each row corresponds to an input gene in `chosen.hvgs`. The batch field in the column metadata contains a vector specifying the batch of origin of each cell.

The `corrected` matrix in the `reducedDims()` contains the low-dimensional corrected coordinates for all cells, which we will use in place of the PCs in our downstream analyses (`r mnn.out.corre.dim[1]` cells and `r mnn.out.corre.dim[2]` PCs).

A `reconstructed` matrix in the `assays()` contains the corrected expression values for each gene in each cell, obtained by projecting the low-dimensional coordinates in corrected back into gene expression space (`r mnn.out.recon.dim[1]` genes and `r mnn.out.recon.dim[2]` cells). We do not recommend using this for anything other than visualization.

The most relevant parameter for tuning `fastMNN()` is `k`, which specifies the number of nearest neighbors to consider when defining MNN pairs. This can be interpreted as the minimum anticipated frequency of any shared cell type or state in each batch. Increasing `k` will generally result in more aggressive merging as the algorithm is more generous in matching subpopulations across batches. It can occasionally be desirable to increase `k` if one clearly sees that the same cell types are not being adequately merged across batches.

```{r fastmnnRbindColData_PBMMC_ETV6-RUNX1}
colDataList <- lapply(rescaled, function(x){colData(x)})
colDataDf <- do.call(rbind, colDataList)
colData(mnn.out) <- DataFrame(colDataDf)
```

## Correction diagnostics

### Mixing between batches

We cluster on the low-dimensional corrected coordinates to obtain a partitioning of the cells that serves as a proxy for the population structure. If the batch effect is successfully corrected, clusters corresponding to shared cell types or states should contain cells from multiple batches. We see that all clusters contain contributions from each batch after correction, consistent with our expectation that the batches are replicates of each other.

```{r fastmnn_clu_PBMMC_ETV6-RUNX1}
snn.gr <- buildSNNGraph(mnn.out, use.dimred="corrected", k=20)
clusters.mnn <- igraph::cluster_walktrap(snn.gr)$membership
colLabels(mnn.out) <- factor(clusters.mnn)
clusterTab.mnn <- data.frame(clusters=clusters.mnn, batch=mnn.out$SampleName, source=mnn.out$SampleGroup)
```

Cluster size and cell contribution by sample, with clusters sorted by size:

```{r fastmnn_diagPlotDrawSampleName_PBMMC_ETV6-RUNX1, fig.height=6, fig.width=7, message=FALSE}
ClusterInfo.mnn <- clusterTab.mnn %>% 
  as.data.frame() %>%
  group_by(clusters,batch) %>%
  summarise(cells = n())

mp1 <- ggplot(data=ClusterInfo.mnn, aes(x=clusters,y=cells, fill=batch)) +
	theme(axis.text.x=element_text(angle = 90, hjust = 0)) +
	geom_col() +
	theme(legend.text = element_text(size = 7))
mp2 <- ggplot(data=clusterTab.mnn, aes(x=clusters, fill=batch)) +
  geom_bar(position = "fill") +
	theme(axis.text.x=element_text(angle = 90, hjust = 0)) +
  scale_y_continuous(labels = scales::percent) +
  theme(legend.text = element_text(size = 7))

plot_grid(mp1, mp2, ncol=1)
```

Cluster size and cell contribution by sample type:

```{r fastmnn_diagPlotDrawSampleType_PBMMC_ETV6-RUNX1, fig.height=6, fig.width=7,message=FALSE}
ClusterInfo.mnn.source <- clusterTab.mnn %>% 
  as.data.frame() %>%
  group_by(clusters,source) %>%
  summarise(cells = n())

mp1.s <- ggplot(data=ClusterInfo.mnn.source, aes(x=clusters,y=cells, fill=source)) +
	theme(axis.text.x=element_text(angle = 90, hjust = 0)) +
	geom_col()
mp2.s <- ggplot(data=clusterTab.mnn, aes(x=clusters, fill=source)) +
  geom_bar(position = "fill") +
	theme(axis.text.x=element_text(angle = 90, hjust = 0)) +
  scale_y_continuous(labels = scales::percent)

plot_grid(mp1.s, mp2.s, ncol=1)
```

We can also compute the variation in the log-abundances to rank the clusters with the greatest variability in their proportional abundances across batches. We can then focus on batch-specific clusters that may be indicative of incomplete batch correction. Obviously, though, this diagnostic is subject to interpretation as the same outcome can be caused by batch-specific populations; some prior knowledge about the biological context is necessary to distinguish between these two possibilities. The table below shows the number of cells for each cluster (row) and sample (column) together with the variance in cell number across these samples ('var' column). 

Also bear in mind that the variance is computed across 7 samples here and only serves to sort clusters.

```{r fastmnn_diagVarComp_PBMMC_ETV6-RUNX1}
tab.mnn <- table(clusters=clusterTab.mnn$clusters, sampleName=clusterTab.mnn$batch)

# Using a large pseudo.count to avoid unnecessarily
# large variances when the counts are low.
norm <- normalizeCounts(tab.mnn, pseudo_count=10)

# Ranking clusters by the largest variances.
rv <- rowVars(norm) %>% 
  round(2)

# show
DataFrame(tab.mnn, var=rv)[order(rv, decreasing=TRUE),]
```

We can also visualize the corrected coordinates using a t-SNE plot. The presence of visual clusters containing cells from both batches provides a comforting illusion that the correction was successful.

```{r fastmnn_diagTsnePlot1_PBMMC_ETV6-RUNX1}
set.seed(0010101010)
mnn.out <- runTSNE(mnn.out, dimred="corrected")
mnn.out <- runUMAP(mnn.out, dimred="corrected")

```

### Challenge 1

Draw the TSNE for the fastmnn corrected data. Compare it with the uncorrected TSNE, what do you think?

Find the challenge markdown for this section in the course materials folder.


### Proportion of Variance

For `fastMNN()`, one useful diagnostic is the proportion of variance within each batch that is lost during MNN correction. Specifically, this refers to the within-batch variance that is removed during orthogonalization with respect to the average correction vector at each merge step. This is returned via the `lost.var` field in the metadata of `mnn.out`, which contains a matrix of the variance lost in each batch (column) at each merge step (row).


```{r fastmnn_diagLostVar_PBMMC_ETV6-RUNX1}
round(metadata(mnn.out)$merge.info$lost.var,2)
```

Large proportions of lost variance (>10%) suggest that correction is removing genuine biological heterogeneity. This would occur due to violations of the assumption of orthogonality between the batch effect and the biological subspace (Haghverdi et al. 2018). In this case, the proportion of lost variance is small, indicating that non-orthogonality is not a major concern.

The following t-SNE shows the clusters identified:

```{r fastmnn_diagTsneClu_PBMMC_ETV6-RUNX1}
mp.tsne.cluster <- plotTSNE(mnn.out, colour_by="label", shape_by="SampleGroup")
mp.tsne.cluster

```

```{r fastmnn_diagTsneCluBySampleName_dsi_PBMMC_ETV6-RUNX1}
mp.tsne.cluster + facet_wrap(~colData(mnn.out)$SampleName)
```

The following t-SNE plots show expression levels of known cell type marker genes.

```{r fastmnn_diagTsneCellTypeMarker_PBMMC_ETV6-RUNX1}
genesToShow <- c(
		 "CD79A", # CD79A 	B ***
		 "CST3", # CST3 	monocytes ***
		 "CD3D", # CD3D 	 T cells ***
		 "HBA1" # HBA1 	 erythrocytes ***
	 	)

tmpInd <- which(rowData(uncorrected)$Symbol %in% genesToShow)
ensToShow <- rowData(uncorrected)$ID[tmpInd]

#B cells:

genex <- ensToShow[1]
p <- plotTSNE(mnn.out, colour_by = genex, by_exprs_values="reconstructed")
p <- p + ggtitle(
		paste("B cells", genex,
		rowData(uncorrected)[genex,"Symbol"])
	)
pB <- p

#T cells:

genex <- ensToShow[3]
p <- plotTSNE(mnn.out, colour_by = genex, by_exprs_values="reconstructed")
p <- p + ggtitle(
		paste("T cells", genex,
		rowData(uncorrected)[genex,"Symbol"])
	)
pT <- p

#monocytes:

genex <- ensToShow[2]
p <- plotTSNE(mnn.out, colour_by = genex, by_exprs_values="reconstructed")
p <- p + ggtitle(
		paste("monocytes", genex,
		rowData(uncorrected)[genex,"Symbol"])
	)
pM <- p

#erythrocytes:

genex <- ensToShow[4]
p <- plotTSNE(mnn.out, colour_by = genex, by_exprs_values="reconstructed")
p <- p + ggtitle(
		paste("erythrocytes", genex,
		rowData(uncorrected)[genex,"Symbol"])
	)
pE <- p
```

```{r fastmnn_diagTsneCellTypeMarkerShow_dsi_PBMMC_ETV6-RUNX1}
gridExtra::grid.arrange(pB + theme(legend.position="bottom"),
                        pT + theme(legend.position="bottom"),
                        pM + theme(legend.position="bottom"),
                        pE + theme(legend.position="bottom"),
                        ncol=2)
```

Compare to the uncorrected values:

```{r uncorr_diagTsneB_PBMMC_ETV6-RUNX1}
# B cells
genex <- ensToShow[1]
p <- plotTSNE(uncorrected, colour_by = genex)
p <- p + ggtitle(
		paste("B cells", genex,
		rowData(uncorrected)[genex,"Symbol"])
	)
pBu <- p

#Compare to the uncorrected values, T cells:

genex <- ensToShow[3]
p <- plotTSNE(uncorrected, colour_by = genex)
p <- p + ggtitle(
		paste("T cells", genex,
		rowData(uncorrected)[genex,"Symbol"])
	)
#print(p)
pTu <- p
#```

#Compare to the uncorrected values, monocytes:

genex <- ensToShow[2]
p <- plotTSNE(uncorrected, colour_by = genex)
p <- p + ggtitle(
		paste("monocytes", genex,
		rowData(uncorrected)[genex,"Symbol"])
	)
pMu <- p

#Compare to the uncorrected values, erythrocytes:

genex <- ensToShow[4]
p <- plotTSNE(uncorrected, colour_by = genex)
p <- p + ggtitle(
		paste("erythrocytes", genex,
		rowData(uncorrected)[genex,"Symbol"])
	)
pEu <- p
```

```{r uncorr_diagTsneCellTypeMarkerShow_PBMMC_ETV6-RUNX1}
gridExtra::grid.arrange(pBu + theme(legend.position="bottom"),
                        pTu + theme(legend.position="bottom"),
                        pMu + theme(legend.position="bottom"),
                        pEu + theme(legend.position="bottom"),
                        ncol=2)
```

### Preserving biological heterogeneity

#### Comparison between within-batch clusters and across-batch clusters obtained after MNN correction

Another useful diagnostic check is to compare the clustering within each batch to the clustering of the merged data. Accurate data integration should preserve variance within each batch as there should be nothing to remove between cells in the same batch. This check complements the previously mentioned diagnostics that only focus on the removal of differences between batches. Specifically, it protects us against cases where the correction method simply aggregates all cells together, which would achieve perfect mixing but also discard the biological heterogeneity of interest.

Ideally, we should see a many-to-1 mapping where the across-batch clustering is nested inside the within-batch clusterings. This indicates that any within-batch structure was preserved after correction while acknowledging that greater resolution is possible with more cells. In practice, more discrepancies can be expected even when the correction is perfect, due to the existence of closely related clusters that were arbitrarily separated in the within-batch clustering. As a general rule, we can be satisfied with the correction if the vast majority of entries are zero, though this may depend on whether specific clusters of interest are gained or lost.

Here is a demonstration with one of the samples, where each entry is colored according to the number of cells with each pair of labels (before and after correction), on the log10 scale with pseudocounts (+10) for a smoother color transition (so a minimum value of log10(0+10) == 1). 

```{r, heatmaps}
tabE1 <- table(paste("after", clusters.mnn[linear_rescaled$batch=="ETV6-RUNX1_1"]),
               paste("before", colLabels(rescaled[[1]])))
heatE1 <- pheatmap(log10(tabE1 +10), cluster_rows = FALSE, cluster_cols = FALSE,
                   main = "ETV6-RUNX1_1")
heatE1
```

The redistribution of cells from one set of clusters to another, here 'within-batch before' and 'across-batch after' correction, may also be visualized with a clustering tree [clustree](https://cran.r-project.org/package=clustree). See the Extended section linked on the course website.
 
#### Rand index

Finally, we can summarize the agreement between clusterings by computing the **Rand index**. This provides a simple metric that we can use to assess the preservation of variation by different correction methods. Larger rand indices (i.e., closer to 1) are more desirable, though this must be balanced against the ability of each method to actually remove the batch effect.

```{r randIndex_comp_PBMMC_ETV6-RUNX1, warning=FALSE}

ariVec <- vector(mode = "numeric", length = 7)
sampleNames <- names(rescaled)
names(ariVec) <- sampleNames

for (i in 1:7) {
  ariVec[i] <- pairwiseRand(
    ref=as.integer(colLabels(rescaled[[i]])),
    alt=as.integer(clusters.mnn[linear_rescaled$batch==sampleNames[i]]),
    mode="index")
}
ariVec <- round(ariVec,2)
ariVec
```

A sample may show a low Rand index value if cells grouped together in a small cluster before correction are split into distinct clusters after correction because the latter comprise cell populations not observed in that sample but present in other samples.

This would be the case of GSM3872434 with far fewer erythrocytes (grouped in a single cluster) than GSM3872443, in which subtypes can be distinguished.

We can also break down the **adjusted Rand index (ARI)** into per-cluster ratios for more detailed diagnostics. For example, we could see low ratios off the diagonal if distinct clusters in the within-batch clustering were incorrectly aggregated in the merged clustering. Conversely, we might see low ratios on the diagonal if the correction inflated or introduced spurious heterogeneity inside a within-batch cluster.

```{r randIndex_plot_PBMMC_ETV6-RUNX1, warning=F}
# pairwiseRand(), ratio, adjusted
# square numeric matrix is returned with number of rows equal to the number of unique levels in ref.

tabList <- vector(mode = "list", length = 7)
for (i in 1:7) {
  tabList[[i]] <- pairwiseRand(
    ref=as.integer(colLabels(rescaled[[i]])),
    alt=as.integer(clusters.mnn[linear_rescaled$batch==sampleNames[i]])
	)
}
randVal <- unlist(tabList) 

## make breaks from combined range
limits <- c(
  min(randVal, na.rm = TRUE),
  max(randVal, na.rm = TRUE))
limits <- quantile(randVal, probs=c(0.05, 0.95), na.rm = TRUE)

Breaks <- seq(limits[1], limits[2],
              length = 100)

plotList <- vector(mode = "list", length = 7)
for (i in 1:7) {
  plotList[[i]] <- pheatmap(tabList[[i]],
                                 cluster_row=FALSE,
                                 cluster_col=FALSE,
                                 breaks=Breaks,
                                 main=sprintf("%s ratio", sampleNames[i]),
                                 silent=TRUE)
}
grobList <- lapply(plotList, function(x){x[[4]]})
gridExtra::grid.arrange(grobs = grobList,
      ncol=2
)
```

We suggest limiting the use of per-gene corrected values to visualization, e.g., when coloring points on a t-SNE plot by per-cell expression. This can be more aesthetically pleasing than uncorrected expression values that may contain large shifts on the colour scale between cells in different batches. Use of the corrected values in any quantitative procedure should be treated with caution, and should be backed up by similar results from an analysis on the uncorrected values.

```{r consis_mnnoutSaveRds_PBMMC_ETV6-RUNX1, warning=FALSE}
# save the mnn.out object in a file

saveRDS(mnn.out, file="./Robjects/DataIntegration_mnn.out.Rds")
```

## Session information

```{r sessionInfo_dsi_PBMMC_ETV6-RUNX1}
sessionInfo()
```
