==> Markdowns/08_ClusteringPostDsi_exercise2_solutions.Rmd <==
---
title: "Introduction to single-cell RNA-seq analysis"
author: "Stephane Ballereau"
date: "November 2021"
subtitle: "Clustering - exercise"
output:
  html_document:
    toc: yes
    toc_depth: 5
    toc_float: yes
    number_sections: yes
    df_print: paged
    fig_caption: yes
    self_contained: yes
    fig_width: 6
    fig_height: 4
  pdf_document:
    toc: yes
    toc_depth: '5'
---

# Clustering - exercise

Imagine ETV6-RUNX1_4 failed leaving you with three ETV6-RUNX1 replicates,
... but the corrected values remain the same somehow.

Define 'walktrap' and Louvain clusters and compare the outcome.
Maybe check clustering quality using per-cluster modularity, too. 

## load data

```{r}
splSetToGet <- "PBMMC,ETV6-RUNX1"
splSetVec <- unlist(strsplit(splSetToGet, ","))

nbPcToComp <- 50
figSize <- 7
```

```{r clusPostDsi_Caron.knitr_options, echo=FALSE, results="hide", message=FALSE}
# Set some variables:
require(knitr)
opts_chunk$set(error=FALSE, message=FALSE, warning=FALSE, cache=FALSE)
opts_chunk$set(echo = TRUE)
opts_chunk$set(eval = TRUE) 
opts_chunk$set(fig.width=7, fig.height=7)
options(stringsAsFactors = FALSE)
set.seed(123) # for reproducibility
```

```{r  clusPostDsi_Caron.library, warning=FALSE, message=FALSE}
library(tidyverse) # data wrangling and plotting (ggplot2)
library(scater) # scRnaSeq QC
library(scran) # scRnaSeq normalisation
library(bluster) # scRnaSeq clustering
library(dynamicTreeCut) # tree cutting in clustering
library(cluster) # for silhouette
library(igraph) # for graph-based clustering and plotting networks 
#library(reticulate) # for graph-based clustering with leiden
#library(leiden) # for community detection (clustering)
library(pheatmap) # for heatmap plotting
library(patchwork) # to combine plots
fontsize <- theme(axis.text=element_text(size=12), axis.title=element_text(size=16))
```


```{r}
sce <- readRDS("../Robjects/DataIntegration_mnn.out.Rds") # auto.merge

# remove ETV6-RUNX1_4
bcToKeep <- colData(sce) %>%
  data.frame() %>%
  rownames_to_column("Barcode2") %>%
  filter(!SampleName == "ETV6-RUNX1_4") %>%
  pull(Barcode2)
indToKeep <- which(colnames(sce) %in% bcToKeep)
sce <- sce[,indToKeep]

sce$SampleName <- factor(sce$SampleName)
sce <- runTSNE(sce, dimred = "corrected")
```

Build graph:

```{r}
# think 'buildSNNGraph'
snn.gr <- buildSNNGraph(sce, use.dimred="corrected")
```

## Walktrap

Compute 'walktrap' clusters:

```{r}
# identify clusters with walktrap
# default number of steps: 4
cluster.out <- cluster_walktrap(snn.gr)
# store membership
sce$walktrap <- cluster.out$membership
# show cluster sizes:
table(sce$walktrap)
```

Embedding plot:

```{r}
p <- plotTSNE(sce, colour_by="walktrap") + fontsize
p
```

Per-cluster modularity

```{r}
# compute
mod.out <- bluster::pairwiseModularity(snn.gr,
                                       sce$walktrap,
                                       get.weights=TRUE)
# Modularity is proportional to the cluster size,
# so we compute the ratio of the observed to expected weights
# for each cluster pair
ratio <- mod.out$observed/mod.out$expected
lratio <- log10(ratio + 1) # on log scale to improve colour range

# plot
pheatmap(lratio, cluster_rows=FALSE, cluster_cols=FALSE, 
    color=colorRampPalette(c("white", "blue"))(100))


```

## Louvain

Compute Louvain clusters:

```{r}
ig.louvain <- igraph::cluster_louvain(snn.gr)
cl <- ig.louvain$membership
cl <- factor(cl)
# store membership
sce$louvain <- cl
# show cluster sizes:
table(sce$louvain)
```


```{r}
fontsize <- theme(axis.text=element_text(size=12), axis.title=element_text(size=16))
# show clusters on TSNE
p <- plotTSNE(sce, colour_by="louvain") + fontsize
p
```

Per-cluster modularity

```{r}
# compute
mod.out <- bluster::pairwiseModularity(snn.gr,
                                       sce$louvain,
                                       get.weights=TRUE)
# Modularity is proportional to the cluster size,
# so we compute the ratio of the observed to expected weights
# for each cluster pair
ratio <- mod.out$observed/mod.out$expected
lratio <- log10(ratio + 1) # on log scale to improve colour range

# plot
pheatmap(lratio, cluster_rows=FALSE, cluster_cols=FALSE, 
    color=colorRampPalette(c("white", "blue"))(100))


```

## Compare walktrap and louvain clusters

```{r}
tmpTab <- table(sce$walktrap, sce$louvain)
tmpTab
```

```{r}
rownames(tmpTab) = paste("walktrap", rownames(tmpTab), sep = "_")
colnames(tmpTab) = paste("louvain", colnames(tmpTab) , sep = "_")
pheatmap(tmpTab)
```
